% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_joke_categories.R
\name{get_joke_categories}
\alias{get_joke_categories}
\title{Get all available joke categories from ICNDB}
\usage{
get_joke_categories()
}
\value{
A vector of available joke categories
}
\description{
Get all available joke categories from ICNDB
}
\examples{
get_joke_categories()
}
