% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_joke_count.R
\name{get_joke_count}
\alias{get_joke_count}
\title{Get number of available jokes in ICNDB}
\usage{
get_joke_count()
}
\value{
A single integer of available jokes in ICNDB
}
\description{
Get number of available jokes in ICNDB
}
\examples{
get_joke_count()
}
