% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiple_randos.R
\name{get_multiple_randos}
\alias{get_multiple_randos}
\title{Get a bunch of random Chuck Norris jokes.}
\usage{
get_multiple_randos(jokes = 5)
}
\arguments{
\item{jokes}{The number of jokes to fetch.}
}
\value{
A tibble of jokes
}
\description{
Get a bunch of random Chuck Norris jokes.
}
\examples{
get_multiple_randos(10)
}
