% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numbered_joke.R
\name{get_numbered_joke}
\alias{get_numbered_joke}
\title{Get a specific joke by number}
\usage{
get_numbered_joke(joke_number)
}
\arguments{
\item{joke_number}{A single integer}
}
\value{
A single joke
}
\description{
Get a specific joke by number
}
\examples{
get_numbered_joke(17)
}
