% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_joke.R
\name{get_random_joke}
\alias{get_random_joke}
\title{Get a random Chuck Norris joke}
\usage{
get_random_joke()
}
\value{
A single joke
}
\description{
Get a random Chuck Norris joke
}
\examples{
get_random_joke()
}
